Refactor: cal_ldos in pw
