+++ source/module_elecstate/CMakeLists.txt
@@ -32,7 +32,6 @@ list(APPEND objects
     magnetism.cpp
     occupy.cpp
     cal_ux.cpp
-    cal_ldos.cpp
     read_orb.cpp
     cal_nelec_nband.cpp
     read_pseudo.cpp
