+++ source/module_elecstate/elecstate_pw.h
@@ -35,9 +35,6 @@ class ElecStatePW : public ElecState
 
     virtual void cal_tau(const psi::Psi<T, Device>& psi);
 
-    //! calculate local density of states
-    virtual void cal_ldos(const psi::Psi<T, Device>& psi, std::vector<double>& ldos) const;
-
     //! calculate becsum for uspp
     void cal_becsum(const psi::Psi<T, Device>& psi);
 
@@ -49,11 +46,11 @@ class ElecStatePW : public ElecState
     T** rhog = nullptr;     // [Device] [spin][nrxx] rhog
     Real** kin_r = nullptr; // [Device] [spin][nrxx] kin_r
 
+    ModulePW::PW_Basis_K* basis = nullptr;
+
   protected:
 
     ModulePW::PW_Basis* rhopw_smooth = nullptr;
-
-    ModulePW::PW_Basis_K* basis = nullptr;
 
     UnitCell* ucell = nullptr;
