+++ source/module_io/CMakeLists.txt
@@ -3,6 +3,7 @@ list(APPEND objects
     bessel_basis.cpp
     cal_test.cpp
     cal_dos.cpp
+    cal_ldos.cpp
     cif_io.cpp
     write_dos_pw.cpp
     nscf_band.cpp
